tests/Document-getElementsByTagName.html

<!DOCTYPE html>
<title>Document.getElementsByTagName</title>
<script src="testharness.js"></script>
<body>
<div id="log"></div>
<script>
test(function() {
  assert_true(document.getElementsByTagName("html") instanceof NodeList, "NodeList")
  assert_false(document.getElementsByTagName("html") instanceof HTMLCollection, "HTMLCollection")
  assert_true(document.getElementsByTagName("html") !== document.getElementsByTagName("html"), "no caching")

  assert_equals(document.createElementNS("http://www.w3.org/1999/xhtml", "i").localName, "i") // Sanity
  var i = document.body.appendChild(document.createElementNS("http://www.w3.org/1999/xhtml", "I"))
  assert_equals(i.localName, "I")
  assert_equals(i.tagName, "I")
  assert_equals(document.getElementsByTagName("I").length, 0)
  assert_equals(document.getElementsByTagName("i").length, 0)
  assert_equals(document.body.getElementsByTagName("I").length, 0)
  assert_equals(document.body.getElementsByTagName("i").length, 0)

  var t = document.body.appendChild(document.createElementNS("test", "te:st"))
  assert_equals(document.getElementsByTagName("st").length, 0)
  assert_equals(document.getElementsByTagName("te:st")[0], t)

  var h = document.body.appendChild(document.createElementNS("http://www.w3.org/1999/xhtml", "te:st"))
  assert_equals(document.getElementsByTagName("st").length, 0)
  assert_equals(document.getElementsByTagName("TE:ST")[0], h)
})
</script>
